% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrdem_dl.R
\name{hrdem_dl}
\alias{hrdem_dl}
\title{Access and Download Digital Elevation Data}
\source{
Government of Canada. (2025). High Resolution Digital Elevation Model (HRDEM) - CanElevation Series.
\cr https://open.canada.ca/data/en/dataset/957782bf-847c-4644-a757-e383c0057995
}
\usage{
hrdem_dl(x, path = ".", model, pattern)
}
\arguments{
\item{x}{an object of class sf, sfc, or sfg}

\item{path}{a character; the path to write the accessed digital elevation
models to.  The default corresponds to the working directory.}

\item{model}{a character; the desired digital elevation model.  Valid values
include "dsm" or "dtm", representing the digital surface model
or digital terrain model, respectively.}

\item{pattern}{a character; the desired digital elevation model's file type.
Valid values include "tif" or "vrt", representing GeoTIFF or
virtual raster file types, respectively.}
}
\description{
Download digital surface or terrain model data from Canada's High Resolution
Digital Elevation Model (HRDEM), masked by a feature geometry.
}
\examples{
point <-
  sf::st_point(c(-98.35814, 49.35103)) |>
  sf::st_sfc(crs = 4326L) |>
  sf::st_transform(26914L)

buffer <-
  sf::st_buffer(point, 403L)

temp_path <-
  tempdir()

hrdem_dl(buffer, temp_path, "dtm", "vrt")

file_paths <-
  list.files(temp_path,
             pattern = "\\\\.vrt$",
             full.names = TRUE)

rasters <-
  lapply(file_paths,
         terra::rast)

raster <-
  rasters[[2]]

terra::plot(raster)

file.remove(file_paths)
}
