% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrdem_mosaic.R
\name{hrdem_mosaic}
\alias{hrdem_mosaic}
\title{Combine Accessed Digital Elevation Data}
\usage{
hrdem_mosaic(x, method = "bilinear", fun = "mean")
}
\arguments{
\item{x}{a list of SpatRasters}

\item{method}{a character; the method used to estimate resampled cell values.
Can be "near", "bilinear", "cubic", "cublicspline", "lanczos",
"sum", "min", "q1", "med", "q3", "max", "average", "mode", or
"rms".}

\item{fun}{a character; the function used to assign values to overlapping
cells.  Can be "mean", "median", "min", "max", "modal", "sum",
"first", or "last".}
}
\value{
SpatRaster
}
\description{
Combine accessed high resolution digital elevation data into a single
mosaciked raster.  Rasters are resampled to the resolution of the lowest
resolution raster using bilinear interpolation and overlapping cells are
averaged, by default.
}
\examples{
point <-
  sf::st_point(c(-97.79828, 49.97858)) |>
  sf::st_sfc(crs = 4326L) |>
  sf::st_transform(26914L)

buffer <-
  sf::st_buffer(point, 403L)

temp_path <-
  tempdir()

hrdem_dl(buffer, temp_path, "dtm", "vrt")

file_paths <-
  list.files(temp_path,
             pattern = "\\\\.vrt$",
             full.names = TRUE)

rasters <-
  lapply(file_paths,
         terra::rast)

raster <-
  hrdem_mosaic(rasters,
               method = "bilinear",
               fun = "mean")

terra::plot(raster)

file.remove(file_paths)
}
